tayfin-ingestor/tayfin-ingestor-jobs/requirements.txt
typer
PyYAML
python-dotenv
sqlalchemy
psycopg[binary]
pandas
tabulate
httpx
lxml